fix(swarm-abm): 3a auditoría — 5 altos, 9 medios, ~10 bajos + bump 0.4.0
Tercera pasada: 4 lectores por subsistema + 1 auditor adversarial
transversal (cobertura de tests, CI, docs vs código, estado de lo
publicado). Cada hallazgo verificado empíricamente; cada fix con test de
regresión. Ver docs/AUDIT.md, sección "Auditoría — 3ª pasada (2026-07-10)".

Altos:
- Morris sin common random numbers (el espejo de F3): semilla compartida
  por trayectoria vía child_rng; parámetro inerte pasa de mu*=0.5 a 0.
- El .skip(1) de F4 rompía el balance diádico de Sobol' (Owen 2020):
  ahora .skip(n.next_power_of_two()); err del estimador ~3 órdenes menos.
- swarm-py no compilaba desde el rename (E0659 crate vs #[pymodule]).
- Bindings fuera de workspace y CI (causa raíz): job bindings nuevo.
- crates.io 0.3.0 contiene F1-F6 sin bump ni CHANGELOG: workspace a
  0.4.0, CHANGELOG reestructurado con la 0.3.0 real y su advertencia.

Medios: NaN propaga NaN en índices Sobol (antes colapsaba todo a 0.0);
wrap toroidal ya no devuelve width exacto; rustdoc de orden por slot (4
sitios pre-arena); collect_every documentado en from_checkpoint; seed u64
en swarm-wasm (BigInt en el visor); Staged bit-idéntico por los 4 entry
points; checkpoint bajo Simultaneous/Staged; features experiment,serde
sin parallel en CI; golden test de trayectoria completa (también wasm32).

Bajos: radio no finito en for_each_within; auto-inclusión en vecindades
con eje de dimensión 1; # Panics de graph.rs; max_base_level entero;
sigma de Morris muestral (n-1); ejemplos migrados a las primitivas rng;
lib.rs y REPRODUCIBILITY.md actualizados (cinco piezas del checkpoint).

Diferido: GIL en run() de swarm-py (necesita reporters + Send en el
motor, documentado); publicar 0.4.0 y re-correr el Sobol de SIGRID.

Gate: build + clippy --all-features -D warnings + test --workspace
--all-features (36 suites) + fmt + doc en verde; swarm-py y swarm-wasm
compilan de nuevo.

+++ crates/swarm-wasm/src/lib.rs
@@ -35,7 +35,7 @@ impl Schelling {
     /// Crea el modelo sobre una grilla `size × size` toroidal.
     #[wasm_bindgen(constructor)]
     #[must_use]
-    pub fn new(size: usize, density: f64, tolerance: f64, seed: u32) -> Schelling {
+    pub fn new(size: usize, density: f64, tolerance: f64, seed: u64) -> Schelling {
         let cfg = schelling::SchellingConfig {
             width: size,
             height: size,
@@ -43,7 +43,7 @@ impl Schelling {
             tolerance,
         };
         Schelling {
-            sim: Simulation::new(schelling::build(cfg, u64::from(seed)), u64::from(seed)),
+            sim: Simulation::new(schelling::build(cfg, seed), seed),
         }
     }
 
@@ -107,7 +107,7 @@ impl Sir {
     /// Crea el modelo sobre una grilla `size × size` toroidal.
     #[wasm_bindgen(constructor)]
     #[must_use]
-    pub fn new(size: usize, beta: f64, gamma: f64, initial_infected: usize, seed: u32) -> Sir {
+    pub fn new(size: usize, beta: f64, gamma: f64, initial_infected: usize, seed: u64) -> Sir {
         let cfg = sir::SirConfig {
             width: size,
             height: size,
@@ -116,7 +116,7 @@ impl Sir {
             gamma,
         };
         Sir {
-            sim: Simulation::new(sir::build(cfg, u64::from(seed)), u64::from(seed)),
+            sim: Simulation::new(sir::build(cfg, seed), seed),
         }
     }
 
@@ -180,7 +180,7 @@ impl Sugarscape {
     /// Crea el modelo sobre una grilla `size × size` con `n_agents` agentes.
     #[wasm_bindgen(constructor)]
     #[must_use]
-    pub fn new(size: usize, n_agents: usize, growback: u32, seed: u32) -> Sugarscape {
+    pub fn new(size: usize, n_agents: usize, growback: u32, seed: u64) -> Sugarscape {
         let cfg = sugarscape::SugarscapeConfig {
             width: size,
             height: size,
@@ -188,7 +188,7 @@ impl Sugarscape {
             growback,
         };
         Sugarscape {
-            sim: Simulation::new(sugarscape::build(cfg, u64::from(seed)), u64::from(seed))
+            sim: Simulation::new(sugarscape::build(cfg, seed), seed)
                 .with_schedule(Schedule::new(Activation::Random)),
         }
     }
